+++ ModConfig.cs
@@ -3,21 +3,9 @@ using System.ComponentModel;
 using System.Runtime.Serialization;
 using Microsoft.Xna.Framework;
 using Newtonsoft.Json;
-// using Terraria.ModLoader.Config;
+using Terraria.ModLoader.Config;
 
 namespace OriMod {
-  #region REMOVE THESE IN tML v0.11.2.2
-  internal class HeaderAttribute : Attribute { internal HeaderAttribute(string s) { } }
-  internal class LabelAttribute : Attribute { internal LabelAttribute(string s) { } }
-  internal class TooltipAttribute : Attribute { internal TooltipAttribute(string s) { } }
-  internal class OptionStringsAttribute : Attribute { internal OptionStringsAttribute(string[] sArr) { } }
-  internal class RangeAttribute : Attribute { internal RangeAttribute(float min, float max) { } }
-  public enum ConfigScope { ClientSide = 1 }
-  public abstract class ModConfig {
-    public abstract ConfigScope Mode { get; }
-    public abstract void OnLoaded();
-  }
-  #endregion
   [Label("Client Config")]
   public class OriConfigClient1 : ModConfig {
     public override ConfigScope Mode => ConfigScope.ClientSide;
@@ -65,11 +53,11 @@ namespace OriMod {
 
     [Label("Player Color"), Tooltip("The color of your Spirit Guardian.")]
     [DefaultValue(typeof(Color), "210, 255, 255, 255")]
-    public Color PlayerColor = new Color(210, 255, 255, 255);
+    public Color PlayerColor;
     
     [Label("Player Color (Secondary)"), Tooltip("The secondary color of your Spirit Guardian.")]
     [DefaultValue(typeof(Color), "0, 0, 0, 0")]
-    public Color PlayerColorSecondary = new Color(0, 0, 0, 0);
+    public Color PlayerColorSecondary;
 
     public override void OnLoaded() {
       OriMod.ConfigClient = this;
@@ -89,93 +77,93 @@ namespace OriMod {
     
     [Label("Cooldown"), Tooltip("Cooldown of Soul Link, in seconds.")]
     [DefaultValue(30), Range(0f, 150)]
-    public float SoulLinkCooldown => 30;
+    public float SoulLinkCooldown;
 
     [Label("Charge Time"), Tooltip("How long it takes to place a Soul Link, in seconds.")]
     [DefaultValue(0.9f), Range(0.05f, 10)]
-    public float SoulLinkChargeRate => 0.9f;
+    public float SoulLinkChargeRate;
 
     [Label("Respawn Time"), Tooltip("How long after dying it takes to respawn at a Soul Link, in seconds.")]
     [DefaultValue(1.2f), Range(1f, 10)]
-    public float SoulLinkRespawnTime => 1.2f;
+    public float SoulLinkRespawnTime;
 
     [Header("Air Jump")]
 
     [Label("Number of Jumps"), Tooltip("Amount of times the player can jump in the air.")]
     [DefaultValue(2)]
-    public int AirJumpCount = 2;
+    public int AirJumpCount;
     
     [Header("Bash")]
 
     [Label("Cooldown"), Tooltip("Cooldown of Bash, in seconds.")]
     [DefaultValue(0.75f), Range(0f, 20)]
-    public float BashCooldown => 0.75f;
+    public float BashCooldown;
 
     [Label("Strength"), Tooltip("How far Bashing pushes the player and NPCs.")]
     [DefaultValue(15), Range(0f, 50)]
-    public float BashStrength => 15;
+    public float BashStrength;
 
     [Label("Range"), Tooltip("Furthest distance from the player a target can be. 1 tile = 16 units.")]
     [DefaultValue(48), Range(8f, 160)]
-    public float BashRange => 48;
+    public float BashRange;
 
     [Header("Stomp")]
 
     [Label("Cooldown"), Tooltip("Cooldown of Stomp, in seconds.")]
     [DefaultValue(6), Range(0f, 30)]
-    public float StompCooldown => 6;
+    public float StompCooldown;
 
     [Label("Fall Speed"), Tooltip("How fast the player falls while Stomping.")]
     [DefaultValue(28), Range(1f, 100)]
-    public float StompFallSpeed => 28;
+    public float StompFallSpeed;
 
     [Label("Number of Targets"), Tooltip("How many targets Stomp can damage at once.")]
     [DefaultValue(8), Range(1, 32)]
-    public int StompNumTargets = 8;
+    public int StompNumTargets;
 
     [Header("Charge Jump")]
 
     [Label("Cooldown"), Tooltip("Cooldown of Charge Jump, in seconds.")]
     [DefaultValue(6), Range(0f, 30)]
-    public float CJumpCooldown => 6;
+    public float CJumpCooldown;
 
     [Label("Speed"), Tooltip("How fast Charge Jump moves the player.")]
     [DefaultValue(1), Range(0f, 10)]
-    public float CJumpSpeedMultiplier => 1;
+    public float CJumpSpeedMultiplier;
 
     [Header("Wall Charge Jump")]
 
     [Label("Cooldown"), Tooltip("Cooldown of Wall Charge Jump, in seconds.")]
     [DefaultValue(4), Range(0f, 30)]
-    public float WCJumpCooldown => 4;
+    public float WCJumpCooldown;
 
     [Label("Speed"), Tooltip("How fast Wall Charge Jump moves the player.")]
     [DefaultValue(1), Range(0f, 10)]
-    public float WCJumpSpeedMultipler => 1;
+    public float WCJumpSpeedMultipler;
 
     [Label("Max Angle"), Tooltip("How far up or down Wall Charge Jump can be aimed.")]
     [DefaultValue(0.65f), Range(0.25f, (float)(Math.PI / 2))]
-    public float WCJumpMaxAngle => 0.65f;
+    public float WCJumpMaxAngle;
 
     [Header("Dash")]
 
     [Label("Cooldown"), Tooltip("Cooldown of Dash, in seconds.\nNote this is only when a boss is alive.")]
     [DefaultValue(1), Range(0f, 10)]
-    public float DashCooldown => 1;
+    public float DashCooldown;
 
     [Label("Speed"), Tooltip("How fast Dash moves the player.")]
     [DefaultValue(1), Range(0f, 10)]
-    public float DashSpeedMultiplier => 1;
+    public float DashSpeedMultiplier;
 
     [Header("Charge Dash")]
     
     [Label("Cooldown"), Tooltip("Cooldown of Charge Dash, in seconds.\nNote this is only when a boss is alive.")]
     [DefaultValue(3), Range(0f, 30)]
-    public float CDashCooldown => 3;
+    public float CDashCooldown;
 
     [Label("Speed"), Tooltip("How fast Charge Dash moves the player.")]
-    [DefaultValue(1)]
-    public float CDashSpeedMultiplier => 1;
+    [DefaultValue(1), Range(0f, 10)]
+    public float CDashSpeedMultiplier;
 
     [Header("Burrow")]